Added CountVectorizer, WordClouds, and started to resample the dataset.
