+++ BackEndStuff/AI/GYM.csv
@@ -1,0 +1,91 @@
+id,subreddit,body
+eyu2ec,GYM,
+fgvlsrc,GYM,thank help get start week
+fjx7ami,GYM,look like good start good effort plan differenti sub fit base one particular r fit big bias toward lift gym work kind sub too
+fjyy9a5,GYM,r fit fantast wiki https thefit wiki constant ad to
+lxj644,GYM,
+gpnccv9,GYM,amount safeti bar seem enough save trash song
+gpnedzi,GYM,man get point might need someth lift home now
+gpn9ld1,GYM,biggest nightmar sinc build home gym enough bench safeti bar
+gpnf3k9,GYM,hate bar gym notch low move high cant touch chest bar
+gpnddd8,GYM,get acquaint roll shame
+gpnmsg5,GYM,safeti first say
+gpnbya9,GYM,isnt healthi elbow pressur
+gpnahyo,GYM,drama queen much weight enough kill lol
+gpncylx,GYM,yeah agre tbh youv got play tiktok audienc ya know
+gpna4oi,GYM,shake hand tell need know bad gonna fail rep haha
+gpnhuh2,GYM,put someth bench leg lift bit use piec leftov lamin floor
+gpnhn2p,GYM,problem rack home perfect height relat bench def gonna bench day home warm up
+gpndnp5,GYM,shit hurt man worth haha especi your wear belt get stuck
+gpnhydu,GYM,yeah found hard way cant roll shame belt least cant collar dump plate
+gpncb1h,GYM,time probabl yeah month old ive chang form quit sinc wrist arent bent back anymor take pressur joint
+gpnaru5,GYM,yeah your right man let get crush that alway good idea
+gpngrza,GYM,still unpleas trap underneath your exhaust
+gpnd7w5,GYM,yeah regular let lbs rest windpip lift it
+gpnb0y6,GYM,liter second start lift would nope immedi fair play see through
+gpnjdj2,GYM,great idea see anyth around use next time d
+gpnimih,GYM,never brave enough fear take fellow gym goer
+lwx3il,GYM,
+gpka15o,GYM,els read arnold voic
+gpk3toz,GYM,sure abl read arney respons voic mind
+gpk8e59,GYM,gym realli transform mental health month would room day horribl depress self harm month go gym sweat feel good almost everyday feel pretti content right good feel final
+gpkvz7l,GYM,would anyon hit man
+gplp379,GYM,nation treasur
+gpk2vpd,GYM,goat reason
+gpju8fj,GYM,arnold good peopl
+gplmqbe,GYM,love arni
+gpjvfsu,GYM,spend much ped motiv get money worth
+gpnjixn,GYM,dope
+gpkmd4s,GYM,read accent too
+gpnl2jw,GYM,hard to
+gpldqvy,GYM,feel like im go opposit way idk work would differ thing keep busi bore often busi gym work want get home eat dinner pass anyth extra seem forc realli enjoy look good tire look fat fun tough choic lol
+gpn4ptg,GYM,mayb diet sleep schedul find eat right get hour sleep everi night lot energi throughout day despit hard train
+lxe76c,GYM,
+gpnajkh,GYM,nice buzzcut work
+lxlig9,GYM,hello guy im kinda new hit gym need help workout im lbs tri gain muscl mass workout routin look like day chest bicep day back tricep day rest day day leg day shoulder ab day rest day good split could improv better result thank help guy
+gpnl675,GYM,routin much schedul day week list exercis set rep order determin effect altern pick written routin wiki sub tri write routin experi
+lwvt9p,GYM,
+gpjncn2,GYM,noic that impress especi take bw account train powerlift
+gpmerde,GYM,good work work leg drive though even weight plant feet flat drive quad
+gpjvtx0,GYM,great job good spotter too
+gpjzxsk,GYM,welcom club friend
+gpk2uek,GYM,current lb max bench hope hit year congrat bro
+gpjovcj,GYM,went easi good job man
+gpjz55g,GYM,nice man
+gpk1xli,GYM,best feel world welcom plate club
+gpk7bye,GYM,delet
+gpl8u99,GYM,good job bro
+gpjuf6y,GYM,wear jean lol
+gplg68k,GYM,cheer train powerlift u tell horribl back arch lmao
+gplg1p1,GYM,thank get soon
+gplkkm8,GYM,tbh way happier hit and
+gpjygtj,GYM,bruh clear jean
+lxhq8e,GYM,probabl seem like stupid question mani peopl dri scoop preworkout grim unless there particular need use dri scoop mani influenc cough splutter without squirm tast crap
+gpnb4rg,GYM,pre prolli tast horribl tast like crap better tast pre easier dri scoop answer hit faster less water reason
+gpnbctw,GYM,recommend best pre current use myprotein
+gpnbjzo,GYM,myprotien euphem someth would recommend woke af buck fella expens know say buy cheap tool use everyday
+gpnbuzz,GYM,yes well spot myprotein euphem anythimg gainz gotta hit protein goal
+lxhvow,GYM,day push bench press x inclin dumbel press seat overhead press x pec dec later rais x skull crusher x rope push x drop set day pull deadlift x lat pull x barbel row x machin row x face pull x barbel curl x seat inclin dumbel curl x day leg barbel squat x rdl x leg press x stand calf rais x leg extens x hamstr curl x day upper pullup x barbel row x bench press x chest fli x later rais x db curl x tricep push x dropset
+gpn4fl0,GYM,ideal want hit muscl group twice per week hit leg day week use upper lower upper lower routin youll see said often best follow proven programm check wiki sticki post access gym day could tri phul
+lxl7a9,GYM,hello everybodi new determin get back gym give real go skinni alway troubl gain weight build muscl heard eat crucial factor work tri gain muscl decid tri come meal plan issu face current though sinc pretti new time meal guess work full time job drive good part day heard spread meal day pretti good idea abl unless prep food ahead time isnt problem realli plan work earli morn work usual go around day day overal look inform schedul time meal mayb idea easi impact meal could prep take work furthermor plane calori day g protein day thing pay attent make meal far micro macro thank much everyon
+gpnlj69,GYM,way much protein use calcul figur actual need www tdeecalcul net eat much soon result fat gain probabl need figur mani calori eat add get calcul goal last meal time matter overal calori eaten day week matter
+lwz55r,GYM,
+gplohyw,GYM,https extern preview redd wk zz axqstyekwdf liba k qm fnehxii jpgauto webp fed f edd ac e fdcd c aa fc fba vs https static wikia nocooki net trailerpark imag e julian gif revis latest scale width cb
+gpjztf3,GYM,wow incred achiev man congrat
+gpkfj6u,GYM,excel power you
+gpkfqls,GYM,awesom bro keep grow
+gpkgcr3,GYM,insan sometim harder gain weight rather lose it
+gpkk843,GYM,great progress chunki
+gpm4p93,GYM,wanna gain weight howd bro
+gpn7o36,GYM,massiv respect buddi began work year ago reason know hard achiev good work
+gpn93yz,GYM,proud man could give tip hard time gain weight kg lbs
+gpjzvyy,GYM,mani thank
+gpn38o3,GYM,nah time get shred
+gpn37kd,GYM,true everybodi differ
+gpn36ok,GYM,thank buddi
+gpn35vi,GYM,well im sure alreadi know fluff calori etc ill skip bunch lifestyl chang help lot tri grab small meal end day eat regular meal way get consum addit calori without plan diet throughout day time wake youll feel hungri repeat cycl slowli increas quantiti food consum regular meal gain weight tri eat fix time fix locat similar quantiti peopl want gain weight typic dont realiz arent eat enough even though feel like basic schedul help lot best type food consum alway liquid think easi drink instead eat make shake healthi calori dens food like oat milk banana etc snack soft food like yoghurt got familiar sever compound movement gym squat bench press etc goal build muscl get stronger weight set repetit provid object measur progress plan workout care understand exercis matter take recoveri serious walk gym purpos good luck it
+gpnakyj,GYM,look comment answer similar question alreadi
+gplplpt,GYM,https www youtub com watchv bpht fupiq https www youtub com watchv bpht fupiq
+gpn3t99,GYM,help much thank much man
+gplq2ca,GYM,de de ecent
+gpn5kt1,GYM,like word help rememb knowledg useless appli last piec advic would make sure get thing done
