+++ BackEndStuff/AI/videogames.csv
@@ -1,0 +1,50 @@
+id,subreddit,body
+hew0fz,videogames,updat rule reflect current state subreddit work cut level spam self promot spoiler follow effect user post spoiler without mention post titl ban immedi user post link youtub video time month ban anyon continu post low qualiti content temporarili ban place spam video effort get like view pleas celebr videogam reminisc old one get hype futur one laugh cesspool garbag content thank
+ik8coz,videogames,that liter dont use r videogam orchastr raid sub dislik tri raid anoth subreddit behalf r videogam pretti simpl ask anyon caught organis carri subreddit raid behalf subreddit receiv day ban dont it
+g3yl2z8,videogames,sorri stupid question im new reddit raid reddit mean s
+g3k9e4h,videogames,know game communiti get angri perceiv childish hobbi outsid idiot communiti stupid thing like raid
+g4rp9qw,videogames,id like add action often get admin involv result site wide ban
+g6wuj1c,videogames,oppress raid r videogam thatll show em
+g8avhax,videogames,sub weak anyon raid know thank sub thank spread raid love rather raid member messag let go raid
+g8cy6ty,videogames,know topic game amaz realli love anticip must play blockbust video game youll never know store action pack gameplay cutscen soooooooo incred jaw drop mind blow https youtu bhjzqb zm q https youtu bhjzqb zm q
+g8des9y,videogames,know topic game amaz anticip blockbust video game must play youll never know store action pack gameplay cutscen soooooooo incred jaw drop mind blow https youtu bhjzqb zm q https youtu nljxcsekrm game extrem fun experi recommend anyon love video game stori amaz control feel smooth charact develop design amaz tri hard game https youtu nljxcsekrm
+g9cfww9,videogames,would effect perman young enough raid theyr young squash e peen let get back game
+gerl1ry,videogames,lmao thought tell peopl particip game raid
+gfegnmf,videogames,mod obvious dont play wow n ofcoz r gay leroy jenkinsss
+ggq1bix,videogames,dont raid man dont ever raid
+gibc61w,videogames,tip iceberg internet less anonym ssome peopl justif forget manner relat note toxic help insult cap wont ever get meaning talk stuff exchang opinion understand includ agre opposit side viewpoint possibl uncivilis toxic environ problem r banvideogam univers hate use kind reason discredit gamer game even straight lie know tough pill swallow let dont dare raid grant right express opinion reddit inc bunch karen abus anyth find tri ban videogam free besid somehow got shit togeth got polit would much bigger threat much bigger necessarili mean signific listen bunch crazi women anyway suppos misogynist rather point peopl shout loudest arent necessarili success
+gmxl9b1,videogames,anyon rememb old xbox game v v multiplay brawl fight game would play rap music ninja would slide bunch charact could jump high build fight onlin multiplay
+g3yl5jb,videogames,group peopl spam anoth subreddit
+g7d0rzv,videogames,glad ask dark too
+gi95bfy,videogames,youtu hmmmmmmmm
+gmnpxjn,videogames,group peopl sub go anoth sub spam how
+gibbplp,videogames,that actual youtub short url use share button youtub com https youtub com video view url alway youtu https youtu be
+gj8ekr2,videogames,make number
+gp1xyoz,videogames,oh damn didnt know thought sort virus link someth
+lxe69h,videogames,
+gpmrtau,videogames,gerudo valley legend zelda oot
+gpmzcpr,videogames,unchart edit bad hear pictur
+gpmorzw,videogames,gobi desert banjo kazooi
+gpmpfkk,videogames,memori dust nier automata
+gpmw499,videogames,shift sand land super mario
+gpn49aq,videogames,agrabah kingdom heart
+gpmtamv,videogames,sinai desert battlefield one
+gpmxzd8,videogames,sand kingdom super mario odyssey
+gpna6ij,videogames,dark cloud k dark chronicl
+gpnfffv,videogames,rare anyon count ice level desert even though make major actual landscap earth dri barren
+gpnkenu,videogames,new super mario bros ds
+lx7ovy,videogames,
+gpm4gqx,videogames,honest obvious fake still make feel good dont ask how
+gpmdbh7,videogames,sorc
+gpn3e49,videogames,made it
+lx06at,videogames,
+gpkh7hv,videogames,say roommat half shave crab ridden pussi high home drug even thursday
+gpm60c5,videogames,get game
+gpkz12m,videogames,ketamin
+gpltpx0,videogames,everyon know cant overdos ketamin silli youll go nappi nap long happen
+gplt67k,videogames,son bitch
+lxljua,videogames,
+lxm20u,videogames,
+lxlpir,videogames,hello everyon school research project film videogam made survey anyon would want answer would great appreci long dont worri survey https doc googl com form e faipqlset qbmegph gy tlgbasvnsvadjnfz fn xubbpk wsiq viewformusp sf link
+lxfy1z,videogames,
+lxl3g0,videogames,
